app/requirements.txt
Flask==2.3.2
librosa==0.10.0.post2
ffmpeg-python==0.2.0
gunicorn==20.1.0

